vmm: Serialize vcpu TLS initialization
Have the builder thread wait for the initialization of the TLS for each
vCPU worker thread. Otherwise, the VcpuEvent::Resume event, which needs
to be processed on the vCPU local thread, may arrive before the TLS has
been initialized, causing a SIGSEGV.

This issue can be easily reproduced when configuring VMs with 8 or more
vCPUs.

Signed-off-by: Sergio Lopez <slp@redhat.com>