vsock: make accept() and explicit operation
So far, when an incoming connection to a listening port was accepted,
the TsiAcceptRsp was automatically injected into the queue without
waiting for a TsiAcceptReq. This may cause a TsiAcceptRsp to get lost
and don't reach the guest. This behavior was easy to reproduce by
generating a significant number of connections to a program listening
to some port in the guest, such as using "wrk" in the host against an
"nginx" instance in the guest.

This commit changes the behavior to require the incoming connection to
be explicitly accepted by sending a TsiAcceptReq message.

This requires a breaking ABI change in libkrunfw, so we also push
KRUNFW_MIN_VERSION from 2 to 3.

Signed-off-by: Sergio Lopez <slp@redhat.com>