init: ensure "/dev" is properly mounted
So far, we were relying on the kernel to mount "/dev", but the problem
with this approach is that, if the "/dev" directory contains some left
overs from the image, the kernel will refuse to make the mount.

Make sure that "/dev" is mounted by doing so in init.c too, ignoring
EBUSY in case the kernel has already mounted it.

Signed-off-by: Sergio Lopez <slp@redhat.com>