fs/macos: avoid deadlock in open_inode
Locks taken in the statement of "match" are held for the duration of
its context. This leads to a deadlock in an unsual path of
open_inode() with self.path_cache lock being held while calling
self.get_inode().

Split taking the lock and matching get_path()'s returned value in two.

Signed-off-by: Sergio Lopez <slp@sinrega.org>