init: don't build init.c on SEV flavor
Due to the dependencies required for SNP support, building init.c with
SEV support is quite cumbersome due to the number of static libraries
required. Since init/init is not bundled inside the virtio-fs device in
the SEV flavor (which doesn't even have a virtio-fs device), in favor of
bundling it in libkrunfw, avoid building it to simplify the process and
avoid confusion.

Fixes #139

Signed-off-by: Sergio Lopez <slp@redhat.com>