Initial port/implmenentation of virtio-net device using passt
For now it is required that passt is running with a bound socket at /tmp/passt_1.socket
Port forwarding is not yet implemented.

When you enter the vm you have to set the IP address of eth0.
To do that, you can use the following command inside the vm:
/usr/sbin/dhclient -r && /usr/sbin/dhclient eth0 -v

Note that for now you have to set set these parameters:
$ sudo sysctl -w net.core.rmem_max=$((16 * 1024 * 1024))
$ sudo sysctl -w net.core.wmem_max=$((16 * 1024 * 1024))

To work around a bug in passt: https://bugs.passt.top/show_bug.cgi?id=74

Signed-off-by: Matej Hrica <mhrica@redhat.com>