Introduce the "efi" feature
The "efi" feature will enable us to produce an EFI flavor the library,
that will be targeted towards running full VMs using EDK2 as firmware.
This was we sacrifice the ultra-fast boot time of the regular flavor in
favor of being able to boot regular disk images, with their own kernel
and initrd, as long they have an EFI layout.

This commit introduces just lays the foundations for this feature. Don't
expect EDK2 to boot properly just from this commit.

Signed-off-by: Sergio Lopez <slp@redhat.com>