hvf: ensure vcpus run in the right thread
For correctness and to support IPIs in future GICv3 implementations, we
need to ensure each HVF vCPU runs in the right thread, so their
respective IDs match each other.

We can't supply an ID to hv_vcpu_create. Instead, it'll return an
incremental integer each time we call it. In order to have deterministic
IDs, we need to serialize each call.

As we already have a Sender for notifying each vCPU TLS initialization,
let's reuse and simply deplay writting to it until hv_vcpu_create has
been called.

While there, also set MPIDR to the same vCPU ID, as this will be needed
to properly support the in-kernel HVF GICv3 implementation.

Signed-off-by: Sergio Lopez <slp@redhat.com>