virtio/fs: Only change credentials if possible
When run in muvm as an unprivileged user, write operations right now
kind of work by accident, since libkrun assumes it runs as root:

- If the access is in root context, it assumes it does not need to
  change user. The access is then made as the user.
- If the access is as the muvm user, it tries to change uid to that
  user, which succeeds as a no-op. The access is made as the user, and
  then it tries to change back to root after it, which spams an error
  message.
- If the access is as any other user, it tries to change to that uid,
  which fails and the access fails.

This behavior, except for the spammy error, is essentially what we want.
Re-do the credential change logic to explicitly implement the above,
without trying to change uid/gid at all if we do not have the
capability.

Signed-off-by: Asahi Lina <lina@asahilina.net>