init: set up code style and format files
While we use rustfmt and clippy for Rust code, we never settled on a
coding style for the C portions of the project (the init binary). This
has caused the files to divert in style, and in some cases to have
multiple styles mixed in.

Let's settle on a coding style, based on LLVM with a few tweaks to make
it reasonably close to what we were doing so far. We do this by adding
a clang-format configuration file and formatting existing files.

The next commit will extend the tests to ensure PRs adhere to the coding
style in clang-format.

Signed-off-by: Sergio Lopez <slp@redhat.com>