devices/legacy: add support for KVM GICv2
In the past, we had support for both requesting KVM the instantiation of
a GICv3 or a GICv2. We lost this ability with the interrupt controller
refactoring, supporting only GICv2.

Turns out the SoC on some aarch64 boards (namely, the Raspberry Pi
family) doesn't implement an architected vGIC, which is required by the
GICv3 implementation in the KVM.

As RPi boards are very popular, and the maintenance cost for this device
is very small, let's reimplement support for requesting KVM the
instantiation of a GICv2.

Signed-off-by: Sergio Lopez <slp@redhat.com>