virtio-block: enable relaxed sync by default on macOS
On macOS, std::fs::File::flush() uses F_FULLFSYNC, which syncs the
buffers to the drive and asks the drive to wait for flushing all the
data to the permanent storage. In contrast, Apple
Virtualization.Framework only issues an fsync(), which doesn't ask the
drive to flush its buffered data.

When on macOS, tell imago to use relaxed sync to match the behavior of
Virtualization.Framework by default.

Related to containers/podman#27216

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>